requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.2
pandas==2.2.0
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
gunicorn==21.2.0
python-dotenv==1.1.1
PyJWT==2.8.0
cryptography==41.0.7
stripe==8.0.0
resend==2.19.0
